Add a api call to return 21 days of snow free days
That is, get the last day that snowed and return the next 21 days as an
array of dates. Will be used to highlight the next 21 days in the
calendar to see visually what days it can't snow. The client could
hightlight the past days that didn't snow green and the future days
yellow or something.